requirements.txt
python-telegram-bot==13.15
python-dotenv>=0.19.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
APScheduler==3.10.4
cachetools==5.3.2
googleapis-common-protos==1.67.0
urllib3==1.26.20
